+++ handler/transport/transport_test.go
@@ -12,7 +12,7 @@ func Test_parseDuration(t *testing.T) {
 	log := logrus.New()
 
 	var target time.Duration
-	parseDuration("1ms", &target, log.WithContext(context.TODO()))
+	parseDuration("1ms", &target, "my-timeout", log.WithContext(context.TODO()))
 
 	if target != 1000000 {
 		t.Errorf("Unexpected duration given: %#v", target)